+++ packages/quirk/lib/quirk.dart
@@ -3,5 +3,4 @@ library;
 
 export 'src/errors.dart';
 export 'src/iterable_extension.dart';
-export 'src/set_extension.dart';
 export 'src/string_extension.dart';